Move classes out of utils/ into the packages that use them
