+++ sdk/scheduler/src/main/java/org/apache/mesos/offer/ValueUtils.java
@@ -4,7 +4,6 @@ import org.apache.mesos.Protos.Resource;
 import org.apache.mesos.Protos.Value;
 import org.apache.mesos.Protos.Value.Range;
 import org.apache.mesos.Protos.Value.Type;
-import org.apache.mesos.util.Algorithms;
 
 import java.util.Collections;
 import java.util.List;
@@ -58,7 +57,7 @@ public class ValueUtils {
     private static Value.Ranges add(Value.Ranges ranges1, Value.Ranges ranges2) {
         List<Range> list1 = ranges1.getRangeList();
         List<Range> list2 = ranges2.getRangeList();
-        List<Range> list = Algorithms.mergeRanges(list1, list2);
+        List<Range> list = RangeAlgorithms.mergeRanges(list1, list2);
         return Value.Ranges.newBuilder().addAllRange(list).build();
     }
 
@@ -90,7 +89,7 @@ public class ValueUtils {
     private static Value.Ranges subtract(Value.Ranges ranges1, Value.Ranges ranges2) {
         List<Range> list1 = ranges1.getRangeList();
         List<Range> list2 = ranges2.getRangeList();
-        List<Range> list = Algorithms.subtractRanges(list1, list2);
+        List<Range> list = RangeAlgorithms.subtractRanges(list1, list2);
         return Value.Ranges.newBuilder().addAllRange(list).build();
     }
 
@@ -133,9 +132,9 @@ public class ValueUtils {
         List<Range> list1 = ranges1.getRangeList();
         List<Range> list2 = ranges2.getRangeList();
 
-        if (Algorithms.rangesEqual(list1, list2)) {
+        if (RangeAlgorithms.rangesEqual(list1, list2)) {
             return 0;
-        } else if (Algorithms.subtractRanges(list1, list2).size() == 0) {
+        } else if (RangeAlgorithms.subtractRanges(list1, list2).size() == 0) {
             return -1;
         } else {
             return 1;
