+++ sdk/scheduler/src/main/java/org/apache/mesos/specification/DefaultCommandSpec.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
-import org.apache.mesos.util.ValidationUtils;
+import org.apache.mesos.specification.validation.ValidationUtils;
 
 import javax.validation.constraints.NotNull;
 import java.net.URI;
