+++ sdk/scheduler/src/main/java/org/apache/mesos/specification/DefaultVolumeSpecification.java
@@ -6,7 +6,7 @@ import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
 import org.apache.mesos.Protos;
-import org.apache.mesos.util.ValidationUtils;
+import org.apache.mesos.specification.validation.ValidationUtils;
 
 /**
  * This class provides a default implementation of the VolumeSpecification interface.
