+++ sdk/scheduler/src/main/java/org/apache/mesos/specification/ReplacementFailurePolicy.java
@@ -3,7 +3,7 @@ package org.apache.mesos.specification;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
-import org.apache.mesos.util.ValidationUtils;
+import org.apache.mesos.specification.validation.ValidationUtils;
 
 import javax.validation.constraints.Min;
