+++ sdk/scheduler/src/main/java/org/apache/mesos/specification/validation/ValidationUtils.java
@@ -1,13 +1,18 @@
-package org.apache.mesos.util;
+package org.apache.mesos.specification.validation;
 
 import javax.validation.*;
 import java.util.Objects;
 import java.util.Set;
 
 /**
- * Various utilities for Validation.
+ * Various utilities for spec validation.
  */
 public class ValidationUtils {
+
+    private ValidationUtils() {
+        // do not instantiate
+    }
+
     public static <T> void validate(T object) {
         ValidatorFactory validatorFactory = Validation.buildDefaultValidatorFactory();
         Validator validator = validatorFactory.getValidator();
