+++ sdk/scheduler/src/main/java/org/apache/mesos/specification/yaml/WriteOnceLinkedHashMap.java
@@ -1,4 +1,4 @@
-package org.apache.mesos.util;
+package org.apache.mesos.specification.yaml;
 
 import javax.validation.ConstraintViolation;
 import javax.validation.ConstraintViolationException;
